Split public test utils from internal tests so we don't get test libraries in build output
